+++ shared/src/models/radiogram/status/radiogram-unread-status.ts
@@ -1,3 +1,4 @@
+import { IsInt, Min } from 'class-validator';
 import { IsValue } from '../../../utils/validators';
 import { getCreate } from '../../utils';
 import type { RadiogramStatus } from './radiogram-status';
@@ -6,11 +7,17 @@ export class RadiogramUnreadStatus implements RadiogramStatus {
     @IsValue('unreadRadiogramStatus')
     public readonly type = 'unreadRadiogramStatus';
 
+    @IsInt()
+    @Min(0)
+    public readonly publishTime: number;
+
     /**
      * @deprecated Use {@link create} instead.
      */
     // eslint-disable-next-line @typescript-eslint/no-useless-constructor, @typescript-eslint/no-empty-function
-    constructor() {}
+    constructor(publishTime: number) {
+        this.publishTime = publishTime;
+    }
 
     static readonly create = getCreate(this);
 }