+++ shared/src/models/utils/spatial-tree.ts
@@ -163,7 +163,7 @@ interface PointRBushElement {
  * @see https://github.com/mourner/rbush#data-format
  */
 class PointRBush extends RBush<PointRBushElement> {
-    toBBox(element: PointRBushElement) {
+    override toBBox(element: PointRBushElement) {
         return {
             minX: element.position.x,
             minY: element.position.y,
@@ -171,10 +171,10 @@ class PointRBush extends RBush<PointRBushElement> {
             maxY: element.position.y,
         };
     }
-    compareMinX(a: PointRBushElement, b: PointRBushElement) {
+    override compareMinX(a: PointRBushElement, b: PointRBushElement) {
         return a.position.x - b.position.x;
     }
-    compareMinY(a: PointRBushElement, b: PointRBushElement) {
+    override compareMinY(a: PointRBushElement, b: PointRBushElement) {
         return a.position.y - b.position.y;
     }
 }